almost working site small fixes left

+++ src/main/java/com/example/webapp/Repositorys/ProjectLinksRepository.java
@@ -1,7 +1,8 @@
 package com.example.webapp.Repositorys;
 
 import com.example.webapp.DatabseClasses.ProjectsLinks;
+import com.mysql.cj.protocol.InternalTime;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ProjectLinksRepository extends CrudRepository<ProjectsLinks,Long> {
+public interface ProjectLinksRepository extends CrudRepository<ProjectsLinks,Integer> {
 }
